requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-admin
flask-wtf
flask-login
dynaconf
email_validator
gunicorn
